requirements.txt
# this requirements is used for deploying in heroku platform for detected python application.
asgiref==3.6.0
astroid==2.15.4
dill==0.3.6
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
install==1.3.5
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.5.1
PyJWT==1.7.1
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
whitenoise==6.4.0
wrapt==1.15.0
